% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEcAn.OSR-package.R
\docType{package}
\name{PEcAn.OSR}
\alias{PEcAn.OSR}
\alias{PEcAn.OSR-package}
\title{PEcAn.OSR: Use PEcAn to inform and manage suites of OpenSimRoot models}
\description{
Tools to connect the structural-functional plant root model OpenSimRoot
(Postma et al 2017; doi:10.1111/nph.14641) to the ecoinformatic workflow
management and data assimilation tool PEcAn
(LeBauer et a 2013; doi:10.1890/12-0137.1).
}
\seealso{
\url{https://rootmodels.gitlab.io},
\url{http:pecanproject.org}
}
