% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.OSR.R
\name{write.config.OSR}
\alias{write.config.OSR}
\title{Writes a configuration file and run scripts for OpenSimRoot}
\usage{
write.config.OSR(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of PFTs to be simulated.
each PFT must have an element named \code{path}}

\item{trait.values}{list or dataframe of trait values to use for this run.
Traits must be named according to PEcAn standard and will be converted
internally to OSR variable names.}

\item{settings}{settings from PEcAn settings file}

\item{run.id}{integer; a unique identifier for the run.}
}
\value{
path to the directory where configurations were saved
(TODO: Is this actually a useful value to return?)
}
\description{
This function edits the specified set of OpenSimRoot default parameters by
updating traits to the values specified in \code{trait.samples},
then writes the result to a single set of configuration files for one
invocation of OpenSimRoot, e.g. one ensemble member.
Since most model runs involve multiple ensemble members, \code{write.config.OSR}
will usually be called by \code{PEcAn.uncertainty::write.ensemble.configs}
rather than invoked directly by the user.
}
