% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jobsh.R
\name{write_aci_jobsh}
\alias{write_aci_jobsh}
\title{Write a Torque script to run OpenSimRoot on the PSU computing cluster}
\usage{
write_aci_jobsh(rundir, outdir, infile, bin = "~/bin/OpenSimRoot")
}
\arguments{
\item{rundir}{path to directory where run will happen}

\item{outdir}{path to which PEcAn-relevant output should be copied after run.
Not used yet -- haven't decided what to copy}

\item{infile}{path to XML OpenSimRoot input file}

\item{bin}{path to OpenSimRoot binary}
}
\value{
nothing; writes file as side effect
}
\description{
Produces a script that submits one single model as its own job.
This means if you have a batch of models to run,
you'll have to submit each one separately,
which is (1) a pain, and (2) kind of defeats the point of a batch job.
TODO: improve this.
}
\keyword{internal}
