% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jobsh.R
\name{write_jobsh}
\alias{write_jobsh}
\title{Write a simple job.sh for local OpenSimRoot execution}
\usage{
write_jobsh(rundir, outdir, infile, bin = "~/bin/OpenSimRoot")
}
\arguments{
\item{rundir}{path to directory where run will happen}

\item{outdir}{path to which PEcAn-relevant output should be copied after run.
Not used yet -- haven't decided what to copy}

\item{infile}{path to XML OpenSimRoot input file}

\item{bin}{path to OpenSimRoot binary}
}
\value{
nothing; writes file as side effect
}
\description{
Write a simple job.sh for local OpenSimRoot execution
}
\examples{
\dontrun{
   osrdir = file.path(tempdir(), "demo")
   write_jobsh(
     file.path(osrdir, "run"),
     file.path(osrdir, "out"),
     infile = "./example.xml",
     binpath = "/usr/bin/OpenSimRoot")
 }
}
\keyword{internal}
